Unbreak cloning from SSH repos with non-default port
remove Repository(); the usage of path.Join() lead to the malformed
URL error as the double slash part of the URI was changed to a
single slash. To add insult to injury, because of that the
Repository() function was added to strip off the 'ssh://' part which
was falsely assumed to be the source of the error. However in doing
that libgit2 went bonkers and we'd end up not being able to clone
repositories on non-default SSH ports (e.g. 2222).